refactor: extract Explorer.expandRecursively function into separate module
